+++ packages/react-native/src/Link.tsx
@@ -33,6 +33,8 @@ export interface LinkState {
 }
 
 class BaseLink extends React.PureComponent<BaseLinkProps, LinkState> {
+  removed: boolean;
+
   state = {
     navigating: false
   };
@@ -53,9 +55,13 @@ class BaseLink extends React.PureComponent<BaseLinkProps, LinkState> {
       let cancelled, finished;
       if (typeof children === "function") {
         cancelled = finished = () => {
-          this.setState({ navigating: false });
+          if (!this.removed) {
+            this.setState({ navigating: false });
+          }
         };
-        this.setState({ navigating: true });
+        if (!this.removed) {
+          this.setState({ navigating: true });
+        }
       }
       router.navigate({
         name,
@@ -90,10 +96,14 @@ class BaseLink extends React.PureComponent<BaseLinkProps, LinkState> {
     return (
       <Anchor {...rest} onPress={this.pressHandler} ref={forwardedRef}>
         {typeof children === "function"
-          ? children(this.state.navigating)
+          ? (children as NavigatingChildren)(this.state.navigating)
           : children}
       </Anchor>
     );
+  }
+
+  componentWillUnmount() {
+    this.removed = true;
   }
 }